[YAM-14] Add action to navigate the last train from HomeScreen. Update icons and add previews

+++ app/src/main/java/com/sedilant/yambol/domain/DomainMapper.kt
@@ -43,7 +43,7 @@ fun AbilityRecordWithName.mapToDomain(): AbilityDomainModel {
 
 fun TrainTaskEntity.mapToDomain(): TrainTaskDomainModel {
     return TrainTaskDomainModel(
-        trainingTaskId = trainingTaskId,
+        trainingTaskId = trainingTaskId.toInt(),
         name = name,
         numberOfPlayer = numberOfPlayers,
         concept = concept,
